requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
distlib==0.3.7
Django==4.2.7
django-filter==23.3
djangorestframework==3.14.0
filelock==3.13.1
gunicorn==21.2.0
packaging==23.2
Pillow==10.1.0
pkg_resources==0.0.0
platformdirs==3.11.0
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.8.0
virtualenv==20.24.6
